+++ frontend/src/App.tsx
@@ -2,7 +2,7 @@ import { BrowserRouter } from 'react-router-dom'
 import './App.css'
 import { LanguageProvider } from './providers/LanguageProvider'
 import CustomRoutes from './routes/CustomRoutes'
-import { AuthProvider } from './contexts/AuthContext'
+import { AuthProvider } from './providers/AuthProvider'
 
 function App() {
