+++ frontend/src/locales/en.json
@@ -1,20 +1,48 @@
 {
-  "alreadyHaveAccount": "Already have an account ?",
+  "alreadyHaveAccount": "You remember your password ?",
   "attemptsRemaining": "Attempts remaining :",
   "backToLogin": "Go back to login page",
+  "ConfirmEmail": "Email confirmation",
   "confirmPassword": "Confirm Password",
+  "CYNA": "CYNA.",
   "email": "Email",
+  "EmailConfirmed": "Email confirmed",
+  "ErrorConfirmationFailed": "Confirmation failed",
+  "ErrorLoggingIn": "Error logging in",
+  "ErrorTokenMissing": "Registration error",
   "ErrorVerifying2FA": "Error verifying 2FA",
   "forgottenPassword": "Password forgotten ?",
   "fullName": "Full Name",
+  "invalidLink": "Invalid link",
+  "InvalidOrExpiredLink": "Invalid or expired link",
   "loading": "Loading...",
+  "loggingIn": "Logging in...",
   "login": "Log in",
+  "logout": "Log out",
+  "newPassword": "New Password",
   "password": "Password",
+  "PasswordRequirements": "Min 8 characters, 1 uppercase, 1 lowercase, 1 number, 1 special character",
+  "passwordResetError": "Error resetting password",
+  "passwordResetSuccess": "Password reset successfully",
+  "RedirectingToLogin": "Redirecting to login page",
   "register": "Register",
+  "RegistrationError": "Registration error",
+  "RegistrationInProgress": "Registration in progress",
+  "RegistrationSuccessful": "Registration successful! Check your email to confirm your account.",
   "rememberMe": "Remember me",
+  "requestNewLink": "Request new link",
+  "requestNewLinkDescription": "Enter your email address and we will send you a link to reset your password.",
+  "requestNewLinkError": "Error requesting new link",
+  "requestNewLinkSuccess": "Request new link sent successfully",
+  "ResendConfirmation": "Resend confirmation",
+  "resetPassword": "Reset password",
+  "resettingPassword": "Resetting password in progress",
+  "sendingLink": "Sending link...",
+  "sendResetLink": "Send reset link",
   "test": "English test",
   "verificationCode": "Verification Code",
   "verificationCodeSent": "Verification code sent to",
+  "VerificationInProgress": "Verification in progress",
   "verify2FA": "Verify 2FA",
   "verifying2FA": "Verifying 2FA",
   "welcome": "Welcome to Cyna.",
