+++ frontend/src/locales/fr.json
@@ -1,20 +1,48 @@
 {
-  "alreadyHaveAccount": "Vous avez déjà un compte ?",
+  "alreadyHaveAccount": "Vous vous souvenez de votre mot de passe ?",
   "attemptsRemaining": "Tentatives restantes :",
   "backToLogin": "Retour à la page de connexion",
+  "ConfirmEmail": "Confirmation d'email",
   "confirmPassword": "Confirmer le mot de passe",
+  "CYNA": "CYNA.",
   "email": "Email",
+  "EmailConfirmed": "Email confirmé",
+  "ErrorConfirmationFailed": "Échec de la confirmation",
+  "ErrorLoggingIn": "Erreur lors de la connexion",
+  "ErrorTokenMissing": "Erreur d'enregistrement",
   "ErrorVerifying2FA": "Erreur de vérification 2FA",
   "forgottenPassword": "Mot de passe oublié ?",
   "fullName": "Nom complet",
+  "invalidLink": "Lien invalide",
+  "InvalidOrExpiredLink": "Lien invalide ou expiré",
   "loading": "Chargement...",
+  "loggingIn": "Connexion en cours...",
   "login": "Se connecter",
+  "logout": "Déconnexion",
+  "newPassword": "Nouveau mot de passe",
   "password": "Mot de passe",
+  "PasswordRequirements": "Min 8 caractères, 1 majuscule, 1 minuscule, 1 chiffre, 1 caractère spécial",
+  "passwordResetError": "Erreur lors de la réinitialisation du mot de passe",
+  "passwordResetSuccess": "Mot de passe réinitialisé avec succès",
+  "RedirectingToLogin": "Redirection vers la page de connexion",
   "register": "S'inscrire",
+  "RegistrationError": "Erreur lors de l'enregistrement",
+  "RegistrationInProgress": "Enregistrement en cours...",
+  "RegistrationSuccessful": "Enregistrement réussi ! Vérifiez votre email pour confirmer votre compte.",
   "rememberMe": "Se souvenir de moi",
+  "requestNewLink": "Demander un nouveau lien",
+  "requestNewLinkDescription": "Entrez votre adresse email et nous vous enverrons un lien pour réinitialiser votre mot de passe.",
+  "requestNewLinkError": "Erreur lors de la demande de nouveau lien",
+  "requestNewLinkSuccess": "Demande de nouveau lien envoyée avec succès",
+  "ResendConfirmation": "Renvoyer la confirmation",
+  "resetPassword": "Réinitialiser le mot de passe",
+  "resettingPassword": "Réinitialisation du mot de passe en cours",
+  "sendingLink": "Envoi du lien...",
+  "sendResetLink": "Envoyer le lien de réinitialisation",
   "test": "Francais test",
   "verificationCode": "Code de vérification",
   "verificationCodeSent": "Code de vérification envoyé à",
+  "VerificationInProgress": "Vérification en cours",
   "verify2FA": "Vérification 2FA",
   "verifying2FA": "Vérification 2FA en cours",
   "welcome": "Bienvenue sur Cyna.",
