+++ frontend/src/types/interfaces/AuthProviderProps.interface.ts
@@ -1,0 +1,3 @@
+export interface AuthProviderProps {
+  children: React.ReactNode;
+}
