+++ frontend/src/types/interfaces/RegisterFormData.interface.ts
@@ -1,0 +1,6 @@
+export interface RegisterFormData {
+  email: string;
+  password: string;
+  confirmPassword: string;
+  fullName: string;
+}
