+++ frontend/src/types/interfaces/RequestResetFormData.interface.ts
@@ -1,0 +1,3 @@
+export interface RequestResetFormData {
+  email: string;
+}
