+++ frontend/src/types/interfaces/ResetPasswordFormData.interface.ts
@@ -1,0 +1,4 @@
+export interface ResetPasswordFormData {
+  password: string;
+  confirmPassword: string;
+}
