+++ frontend/src/types/interfaces/User.interface.ts
@@ -1,0 +1,7 @@
+export interface User {
+  id: string;
+  email: string;
+  full_name: string;
+  email_verified: boolean;
+  role: 'ADMIN' | 'COMMERCIAL' | 'USER';
+}
