+++ gateway/.env.example
@@ -18,3 +18,7 @@ MS_P_URL=
 LOG_LEVEL=
 
 ALLOWED_ORIGINS=
+
+EMAIL_USER=
+EMAIL_PASSWORD=
+FRONTEND_URL=
