+++ gateway/src/services/mail.service.ts
@@ -1,5 +1,6 @@
 import nodemailer from 'nodemailer';
 import { Logger } from '../common/logger';
+import emailTemplates from '../templates/emailTemplates';
 
 export class MailService {
   private transporter = nodemailer.createTransport({
@@ -13,18 +14,13 @@ export class MailService {
   async sendConfirmationEmail(email: string, token: string) {
     try {
       const confirmUrl = `${process.env.FRONTEND_URL}/confirm-email?token=${token}`;
+      const template = emailTemplates.fr.confirmEmail;
 
       const mailOptions = {
         from: process.env.EMAIL_USER,
         to: email,
-        subject: 'Confirmez votre email - Cyna',
-        html: `
-          <h1>Bienvenue sur Cyna !</h1>
-          <p>Cliquez sur le lien ci-dessous pour confirmer votre email :</p>
-          <a href="${confirmUrl}">Confirmer mon email</a>
-          <p>Ou copiez-collez ce lien : ${confirmUrl}</p>
-          <p>Ce lien expire dans 24h.</p>
-        `,
+        subject: template.subject,
+        html: template.html(confirmUrl),
       };
 
       await this.transporter.sendMail(mailOptions);
@@ -38,18 +34,13 @@ export class MailService {
   async sendPasswordResetEmail(email: string, token: string) {
     try {
       const resetUrl = `${process.env.FRONTEND_URL}/reset-password?token=${token}`;
+      const template = emailTemplates.fr.resetPassword;
 
       const mailOptions = {
         from: process.env.EMAIL_USER,
         to: email,
-        subject: 'Réinitialiser votre mot de passe - Cyna',
-        html: `
-          <h1>Réinitialisation de mot de passe</h1>
-          <p>Cliquez sur le lien ci-dessous pour réinitialiser votre mot de passe :</p>
-          <a href="${resetUrl}">Réinitialiser mon mot de passe</a>
-          <p>Ou copiez-collez ce lien : ${resetUrl}</p>
-          <p>Ce lien expire dans 1h.</p>
-        `,
+        subject: template.subject,
+        html: template.html(resetUrl),
       };
 
       await this.transporter.sendMail(mailOptions);
@@ -62,17 +53,12 @@ export class MailService {
 
   async send2FACode(email: string, code: string) {
     try {
+      const template = emailTemplates.fr.twoFactorCode;
       const mailOptions = {
         from: process.env.EMAIL_USER,
         to: email,
-        subject: 'Votre code de vérification - Cyna',
-        html: `
-          <h1>Code de vérification</h1>
-          <p>Votre code de vérification est :</p>
-          <h2 style="font-size: 32px; letter-spacing: 5px; font-weight: bold;">${code}</h2>
-          <p>Ce code expire dans 5 minutes.</p>
-          <p>Ne partagez ce code avec personne.</p>
-        `,
+        subject: template.subject,
+        html: template.html(code),
       };
 
       await this.transporter.sendMail(mailOptions);
